requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.27.8
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
urllib3==2.0.3
whitenoise==6.5.0
yarl==1.9.2
